src/main/java/com/szymek/socializr/service/PostServiceImpl.java
package com.szymek.socializr.service;

import com.szymek.socializr.model.Post;
import com.szymek.socializr.repository.PostRepository;
import org.springframework.stereotype.Service;

import java.util.Collection;

@Service
public class PostServiceImpl implements PostService{

    private final PostRepository postRepository;

    public PostServiceImpl(PostRepository postRepository) {
        this.postRepository = postRepository;
    }

    @Override
    public Collection<Post> findAllByAuthor() {
        return null;
    }

    @Override
    public Collection<Post> findAll() {
        return (Collection<Post>) postRepository.findAll();
    }

    @Override
    public Post findById(Long postId) {
        return postRepository.findById(postId).get();
    }

    @Override
    public Post create(Post post) {
        return postRepository.save(post);
    }

    @Override
    public void delete(Post post) {
        postRepository.delete(post);
    }

    @Override
    public void deleteById(Long postId) {
        postRepository.deleteById(postId);
    }
}
